Moved subset_to_equal_ratio to utils, and fixed a bug that occurred when the train and val set had exactly the same label distribution
